Moved make meme function to another file
